refactor: move the examples inside the stories and remove unuseful comments
