--- stories/Table/Table.stories.ts
@@ -1,28 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import { TableDemo } from './Table.example'
-
-// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
-const meta = {
-  title: 'Components/Table',
-  component: TableDemo,
-  parameters: {
-    // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
-    layout: 'centered'
-  }
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
-
-  // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
-} satisfies Meta<typeof TableDemo>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-export const Default: Story = {}
-
-export const Zebra: Story = {
-  args: {
-    variant: 'zebra'
-  }
-}
